feat: Add createTask endpoint and Task model
The code changes include adding a new endpoint `createTask` to the `tasks` controller, which allows users to create a new task. It uses the `Task` model to create a task in the database and returns the created task in the response. If there is an error, it sends a 500 status code with an error message.

This commit also includes the addition of the `Task` model, which defines the schema for tasks in the database. It includes fields for `name` (a required string with a maximum length of 20 characters) and `completed` (a boolean with a default value of `false`).

Recent user commits:
- chore: Update mongoose dependency to version 8.5.1
- Add routes and controllers for tasks

Recent repository commits:
- chore: Update mongoose dependency to version 8.5.1
- Add routes and controllers for tasks
- first commit